mcu_gen.py: replace `in` with `==` for string cmp
This is the same issue that was found at bug #548, although in this case
it has not shown any adverse effects *for now*.
Nevertheless, it is never a bad idea to fix it as well just in case.

Also fixed file `pad_cfg.hjson` which incorrectly adds a comma after
quoteless string `active: low,` which isn't valid Hjson.
This worked before because `"low" in "low,"` is technically true,
but this was just one bug hiding another bug.
This issue affects several other files and will be reported separately.