Fix hjson files that shall not use commas (#552)
Some hjson files in X-HEEP use trailing comma after a "quoteless string"
which isn't valid in the standard (https://hjson.github.io/syntax.html):

> Do not add commas or comments as they would become part of the string.

This includes "barewords" such as `active: low,` but also (surprisingly)
hexadecimal values such as `address: 0x10000000,` which is technically
also interpreted as a string and not an integer value.
(`42,`, `true,`, `false,` and `"string",` are OK though.)

This fixes issue #552.
It also deprecates certain workarounds in the code that were using
`.split(',')` or `.strip(',')` to address this issue (but the issue was
really that the file format was wrong, not that the parser failed).